+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 [unreleased]
 * add more Site Health info for runtime environment
-* update `wp_get_runtime_environment()` and `move_dir()`
+* update `move_dir()`
+* add `is_virtualbox()`
+* remove `WP_RUNTIME_ENVIRONMENT` and `wp_get_runtime_environment()`
 
 #### 1.3.4 / 2022-03-21
 * run `restore_temp_backup()` in `shutdown` hook